+++ BankApp2/Bank.cs
@@ -5,7 +5,7 @@ using System.Security.Principal;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BankApp2
+namespace BankApp2.Models
 {
     public class Bank
     {
@@ -41,80 +41,6 @@ namespace BankApp2
                 }
                 
             }
-        }
-       
-                
-            }
-        }
-
-        public void PrintAccounts(User user)
-        {
-            Console.Clear();
-            for (int i = 0; i < user.Account.Count; i++)
-            {
-                Console.WriteLine("-----------------------------");
-                Console.WriteLine($"Account number: {user.Account[i].AccountNumber}");
-                if (user.Account[i] is SavingsAccount)
-                {
-                    Console.WriteLine($"Account type: Savings account");
-                }
-                else if (user.Account[i] is CheckingAccount)
-                {
-                    Console.WriteLine($"Account type: Checking account");
-                }
-                Console.WriteLine($"Balance: {user.Account[i].Balance}");
-                Console.WriteLine("-----------------------------");
-            }
-            Console.ReadKey();
-        }
-
-        public void FindAccount(string id)
-        {
-            foreach (var user in Users)
-            {
-                if (user.id == id)
-                    Console.WriteLine("hittade user");
-
-       //public void FindAccount(string id)
-       //{
-       //    foreach (var user in Users)
-       //    {
-       //        if (user.id == id)
-       //            Console.WriteLine("hittade user");
-       //
-       //        else
-       //        {
-       //            Console.WriteLine("hittade inte user");
-       //        }
-       //    }
-       //}
-                else
-                {
-                    Console.WriteLine("hittade inte user");
-                }
-            }
-        }
-
-        public void PrintPositiveAccounts()
-        {
-            var positiveAccounts = Accounts.Where(a => a.Balance > 0);
-            Console.Clear();
-            Console.WriteLine("Konton med positivt saldo:\n");
-
-            if (positiveAccounts.Count() <= 0)
-            {
-                Console.WriteLine("Inga konton med postivt saldo");
-            }
-            else
-            {
-                foreach (var account in positiveAccounts)
-                {
-                    Console.WriteLine($"Ägare: {account.Owner}");
-                    Console.WriteLine($"Kontonummer: {account.AccountNumber}");
-                    Console.WriteLine($"Saldo: {account.Balance}");
-                }
-            }
-            Console.ReadKey();
         }
     }
 }
