+++ BankApp2/Program.cs
@@ -1,6 +1,7 @@
-﻿using System;
+﻿using BankApp2.Users;
+using System;
 
-namespace BankApp2
+namespace BankApp2.Models
 {
     internal class Program
     {
@@ -44,7 +45,7 @@ namespace BankApp2
                     Console.WriteLine("0. Exit");
                     Console.WriteLine("1. Open new account");
                     Console.WriteLine("2. View accounts");
-                    Console.WriteLine("3. Exit");
+                    Console.WriteLine("3. View accounts with positive balance");
 
                     string response = Console.ReadLine();
                     if (response == "0")
@@ -58,12 +59,12 @@ namespace BankApp2
                         bank.OpenAccount(loginResult.LoggedInUser, random.ToString());
                     }
                     else if (response == "2")
-                        bank.PrintAccounts(loginResult.LoggedInUser);
                     {
+                        loginResult.LoggedInUser.PrintAccounts();
                     }
                     else if (response == "3")
                     {
-                        Environment.Exit(0);
+                        loginResult.LoggedInUser.PrintPositiveAccounts();
                     }
                 }
             }
